chore: Move fux notification to credit balance
