+++ web_ui/src/shared/components/header/credit-balance/fux-notification/fux-notification.module.scss
@@ -1,4 +1,4 @@
-@use '../../shared.module.scss';
+@use '../../../../shared.module';
 
 .container {
     @extend .blueGreenGradient;
